Queue basic manipulation and bug fix
In this commit, we complete the basic handler for queue manipulation
listed in the following
- q_free()
- q_insert_tail()
- q_remove_head()
- q_remove_tail()

Also we had some memory allocation issue related to some functions
like q_new(), q_insert_head() and so on. We change the utilization of
calloc() to malloc() and strndup() to strdup() when allocating memory
for string. Further discussion are needed to explain the differences
between these APIs and why does it cause memory allocation bug.

At the same time we implement a few functions related to leetcode
problems, listed in below.
- q_delete_mid()
- q_delete_dup()
- q_swap()
- q_reverse()