Implementation of Fisher-Yates shuffle Algorithm
Implement the function q_shuffle() based on Fisher-Yates shuffle
algorithm. Shuffle the elements in the queue randomly and generate
random permutation from the original sequence, however, the current
implementation doesn't seems to generate an unbiased permutation.

Future improvements and experiments are needed to make sure the
probability of generation for different permutation is unbiased and
equally distributed.