Correct the traversal of q_shuffle
The original traversal direction of the pointer to struct list_head
`pick` is going to its next pointer. However, the method will cause an
biased result for different permutation.

We changed the traversal direction of `pick` from next to prev.
Statistics experiments are needed to prooved that each permutation is
unbiased.