Re-design fixed-point calculation
Previous version of fixed-point design makes the AI with MCTS suffered
with a poor performance in game. We reconsidered the scaling factor and
re-design the fixed-point for MCTS.

We change the scaling factor from 10 bits to 8 bits which is exactly a
power of two, it'll prevent to discard integer data part from the number
and also achieve a precision with 6 digits in decimal.

Besides changing the scaling factor, we also alter the algorithm to
calculate the square root of a fixed-point number, now we use an
approximation method which can slowly reach the answer.

Future improvements are needed to speed up the process when calculating
logarithm and square root for fixed-point number.