Implementation of co-routine for games between AI
In order to provide concurrency when executing "ai_civil_war" mode in
"do_ttt()", we need to implement coroutine without using pthread in the
project.

Implement coroutine based on "setjmp()" and "longjmp()" so we can
simulate context-switch between different functions in runtime. We'll
store the current function status in a structure called "jmp_buf"