Implement keyboard event watchdog
Implement the function "listen_keyboard()" as the watchdog for keyboard
event. We'll treat "Ctrl+Q" as a signal to stop the tic-tac-toe game
between AI, and treat "Ctrl+P" as a signal to stop or resume display the
chessboard on terminal.

Now the maximum number of available games is 10, future improvements are
needed to extend the number to at least 100 games, or even no
boundaries.