Implement a new pseudo-random number generator
Due to the poor performance of the PRNG we used, which is "mt19937", we
implement a new PRNG to produce pseudo-random number in a faster manner

Implement a new algorithm which is called "wyhash". It's a fast
implementation of pseudo-random number generator which can pass big
crush test and use less than 0.3 second when performing the generation
of pseudo-random number for more than 10^9 times.